# Conflicts:
#	.idea/gradle.xml
#	.idea/modules.xml
#	app/src/main/java/com/yibao/music/activity/SearchLyricsActivity.java
#	app/src/main/java/com/yibao/music/base/BaseLazyFragment.java
#	app/src/main/java/com/yibao/music/fragment/PlayListFragment.java

+++ app/src/main/java/com/yibao/music/fragment/dialogfrag/DeletePlayListDialog.java
@@ -3,19 +3,21 @@ package com.yibao.music.fragment.dialogfrag;
 import android.app.AlertDialog;
 import android.app.Dialog;
 import android.os.Bundle;
-import androidx.annotation.NonNull;
-import androidx.fragment.app.DialogFragment;
 import android.view.Gravity;
 import android.view.View;
 import android.view.Window;
 import android.widget.TextView;
+
+import androidx.annotation.NonNull;
+import androidx.fragment.app.DialogFragment;
+import androidx.swiperefreshlayout.widget.SwipeRefreshLayout;
 
 import com.yibao.music.MusicApplication;
 import com.yibao.music.R;
 import com.yibao.music.model.AddAndDeleteListBean;
 import com.yibao.music.model.PlayListBean;
 import com.yibao.music.model.greendao.MusicBeanDao;
-import com.yibao.music.util.Constants;
+import com.yibao.music.util.Constant;
 import com.yibao.music.util.MusicDaoUtil;
 import com.yibao.music.util.RxBus;
 
@@ -37,8 +39,10 @@ public class DeletePlayListDialog
     private PlayListBean mPlayListBean;
     private RxBus mBus;
     private int mPageType;
+    private static SwipeRefreshLayout.OnRefreshListener mListener;
 
-    public static DeletePlayListDialog newInstance(PlayListBean musicInfo, int pageType) {
+    public static DeletePlayListDialog newInstance(PlayListBean musicInfo, int pageType, SwipeRefreshLayout.OnRefreshListener listener) {
+        mListener = listener;
         Bundle bundle = new Bundle();
         bundle.putParcelable("musicInfo", musicInfo);
         bundle.putInt("pageType", pageType);
@@ -78,7 +82,7 @@ public class DeletePlayListDialog
         mTvCancelDelete = mView.findViewById(R.id.tv_delete_list_cancel);
         mTvDelete = mView.findViewById(R.id.tv_delete_list_continue);
         mBus = RxBus.getInstance();
-        MusicBeanDao musicDao = MusicApplication.getIntstance().getMusicDao();
+        MusicBeanDao musicDao = MusicApplication.getInstance().getMusicDao();
         mPlayListBean = getArguments().getParcelable("musicInfo");
         mPageType = getArguments().getInt("pageType");
         if (mPlayListBean != null) {
@@ -89,26 +93,23 @@ public class DeletePlayListDialog
 
     @Override
     public void onClick(View v) {
-        switch (v.getId()) {
-            case R.id.tv_delete_list_cancel:
-                dismiss();
-                break;
-            case R.id.tv_delete_list_continue:
-                deletePlayList();
-                dismiss();
-                break;
-            default:
-                break;
+        int id = v.getId();
+        if (id == R.id.tv_delete_list_cancel) {
+            dismiss();
+        } else if (id == R.id.tv_delete_list_continue) {
+            deletePlayList();
+            dismiss();
         }
     }
 
 
     private void deletePlayList() {
-        if (mPageType == Constants.NUMBER_TWO) {
+        if (mPageType == Constant.NUMBER_TWO) {
             // 同步更新列表中，的歌曲的列表标识 (更新为“LSP_98”)
             MusicDaoUtil.setMusicListFlag(mPlayListBean);
         }
         mBus.post(new AddAndDeleteListBean(mPageType));
+        mListener.onRefresh();
     }
